+++ server/src/storage/v1/macros.rs
@@ -99,13 +99,6 @@ macro_rules! concat_str {
     }}};
 }
 
-#[macro_export]
-macro_rules! bad_data {
-    () => {
-        std::io::Error::from(std::io::ErrorKind::InvalidData)
-    };
-}
-
 macro_rules! read_dir_to_col {
     ($root:expr) => {
         std::fs::read_dir($root)?