+++ CHANGELOG.md
@@ -9,7 +9,11 @@ All changes in this project will be noted in this file.
 - `INSPECT KEYSPACE` without arguments to inspect the current keyspace
 - `INSPECT TABLE` without arguments to inspect the current table
 - `AUTH WHOAMI` returns the AuthID of the currently logged in user
+
+### Improvements
+
 - Enable multiples values to be pushed into a list at once with `lmod push`
+- (skyd) Improved error reporting infrastructure with more details and context
 
 ### Fixes